+++ src/avt/Database/Formats/avtMTMDFileFormatInterface.C
@@ -349,7 +349,7 @@ void
 avtMTMDFileFormatInterface::SetDatabaseMetaData(avtDatabaseMetaData *md,
     int timeState, bool forceReadAllCyclesTimes)
 {
-    int i;
+    size_t i;
 
     GenerateTimestepCounts();
 
@@ -373,7 +373,7 @@ avtMTMDFileFormatInterface::SetDatabaseMetaData(avtDatabaseMetaData *md,
     int tsGroup = GetTimestepGroupForTimestep(timeState);
     int localTS = GetTimestepWithinGroup(timeState);
     int offset = 0;
-    for (i = 0 ; i < nTimestepGroups ; i++)
+    for (i = 0 ; i < (size_t)nTimestepGroups ; i++)
     {
         chunks[i]->SetTimeSliceOffset(offset);
         offset += tsPerGroup[i];
@@ -383,8 +383,8 @@ avtMTMDFileFormatInterface::SetDatabaseMetaData(avtDatabaseMetaData *md,
     else
         chunks[tsGroup]->SetReadAllCyclesAndTimes(false);
     chunks[tsGroup]->SetDatabaseMetaData(md, localTS);
-    for (i = 0 ; i < nTimestepGroups ; i++)
-        if (i != tsGroup)
+    for (i = 0 ; i < (size_t)nTimestepGroups ; i++)
+        if (i != (size_t)tsGroup)
             chunks[i]->RegisterDatabaseMetaData(md);
 
     //
@@ -395,7 +395,7 @@ avtMTMDFileFormatInterface::SetDatabaseMetaData(avtDatabaseMetaData *md,
     if (md->AreAllCyclesAccurateAndValid(nTotalTimesteps) != true)
     {
         vector<int> cycles;
-        for (i=0; i<nTimestepGroups; i++)
+        for (i=0; i < (size_t)nTimestepGroups; i++)
         {
             vector<int> tmp;
             chunks[i]->FormatGetCycles(tmp);
@@ -413,13 +413,13 @@ avtMTMDFileFormatInterface::SetDatabaseMetaData(avtDatabaseMetaData *md,
                 }
             }
         }
-        if (cycles.size() != nTotalTimesteps)
+        if (cycles.size() != (size_t)nTotalTimesteps)
             cyclesLookGood = false;
         if (cyclesLookGood == false)
         {
             std::vector<int> cyclesFromMassCall = cycles;
             cycles.clear();
-            for (i = 0; i < nTotalTimesteps; i++)
+            for (i = 0; i < (size_t)nTotalTimesteps; i++)
             {
                 int tsg = GetTimestepGroupForTimestep(i);
                 int lts = GetTimestepWithinGroup(i);
@@ -452,7 +452,7 @@ avtMTMDFileFormatInterface::SetDatabaseMetaData(avtDatabaseMetaData *md,
     {
         // Set the times in the metadata.
         vector<double> times;
-        for (i=0; i<nTimestepGroups; i++)
+        for (i=0; i<(size_t)nTimestepGroups; i++)
         {
             vector<double> tmp;
             chunks[i]->FormatGetTimes(tmp);
@@ -470,13 +470,13 @@ avtMTMDFileFormatInterface::SetDatabaseMetaData(avtDatabaseMetaData *md,
                 }
             }
         }
-        if (times.size() != nTotalTimesteps)
+        if (times.size() != (size_t)nTotalTimesteps)
             timesLookGood = false;
         if (timesLookGood == false)
         {
             vector<double> timesFromMassCall = times;
             times.clear();
-            for (i = 0; i < nTotalTimesteps; i++)
+            for (i = 0; i < (size_t)nTotalTimesteps; i++)
             {
                 int tsg = GetTimestepGroupForTimestep(i);
                 int lts = GetTimestepWithinGroup(i);
@@ -548,7 +548,7 @@ avtMTMDFileFormatInterface::SetCycleTimeInDatabaseMetaData(
     bool cyclesLookGood = true;
     if (!isEnsemble)
     {
-        for (int i = 0; i < cycles.size(); i++)
+        for (size_t i = 0; i < cycles.size(); i++)
         {
             if ((i != 0) && (cycles[i] <= cycles[i-1]))
             {
@@ -557,7 +557,7 @@ avtMTMDFileFormatInterface::SetCycleTimeInDatabaseMetaData(
             }
         }
     }
-    if (cycles.size() != nTotalTimesteps)
+    if (cycles.size() != (size_t)nTotalTimesteps)
         cyclesLookGood = false;
     if (cyclesLookGood == false)
     {
@@ -584,7 +584,7 @@ avtMTMDFileFormatInterface::SetCycleTimeInDatabaseMetaData(
         times.insert(times.end(),tmp.begin(),tmp.end());
     }
     bool timesLookGood = true;
-    for (int i = 0; i < times.size(); i++)
+    for (size_t i = 0; i < times.size(); i++)
     {
         if ((i != 0) && (times[i] <= times[i-1]))
         {
@@ -592,7 +592,7 @@ avtMTMDFileFormatInterface::SetCycleTimeInDatabaseMetaData(
             break;
         }
     }
-    if (times.size() != nTotalTimesteps)
+    if (times.size() != (size_t)nTotalTimesteps)
         timesLookGood = false;
     if (timesLookGood == false)
     {
